Plugins: Move constants from state to lib
